🏗️ Pull Tools5eJsonSourceCopier enums into abstract class
